Refactor packages to follow GeoTrellis 3.2.0
