--- src/main/scala/geotrellis/pointcloud/spark/io/s3/S3URIRecordReader.scala
@@ -1,30 +1,0 @@
-/*
- * Copyright 2018 Azavea
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package geotrellis.pointcloud.spark.io.s3
-
-import geotrellis.spark.store.s3.BaseS3RecordReader
-
-import software.amazon.awssdk.services.s3.S3Client
-import software.amazon.awssdk.services.s3.model.GetObjectRequest
-import java.net.URI
-
-abstract class S3URIRecordReader[K, V](s3Client: S3Client) extends BaseS3RecordReader[K, V](s3Client: S3Client) {
-  def readObjectRequest(objectRequest: GetObjectRequest): (K, V) =
-    read(objectRequest.key, new URI(s"s3://${objectRequest.bucket}/${objectRequest.key}"))
-
-  def read(key: String, uri: URI): (K, V)
-}
