+++ src/test/scala/geotrellis/pointcloud/spark/dem/PointCloudDemSpec.scala
@@ -18,7 +18,7 @@ package geotrellis.pointcloud.spark.dem
 
 import geotrellis.raster._
 import geotrellis.pointcloud.spark.buffer._
-import geotrellis.pointcloud.spark.io.hadoop.HadoopPointCloudRDD
+import geotrellis.pointcloud.spark.store.hadoop.HadoopPointCloudRDD
 import geotrellis.pointcloud.spark.{PointCloudTestEnvironment, _}
 import geotrellis.layer._
 import geotrellis.vector.Extent
